created test file

+++ main.py
@@ -1,10 +1,15 @@
 from flask import Flask, request, jsonify
 from pymongo import MongoClient
 from bson import ObjectId
+import os
+
 app = Flask(__name__)
 app.config["DEBUG"] = True
 
-client = MongoClient("mongodb://mongo:27017/")
+#Default to local mongoDB, but docker can override
+MONGO_URI = os.getenv("MONGO_URI", "mongodb://localhost:27017/")
+client = MongoClient(MONGO_URI)
+
 
 db = client["quizAppDB"]
 user_collection = db["users"]
@@ -14,10 +19,18 @@ quiz_collection = db["quizzes"]
 user_collection.create_index([("username", 1)], unique=True)
 quiz_collection.create_index([("quizName", 1)], unique=True)
 
+#Allow the test file to override the main DB
+def create_test_db(testDB):
+    global db, user_collection, quiz_collection
+    db = testDB
+    user_collection = db["users"]
+    quiz_collection = db["quizzes"]
+
+
 #home page
 @app.route('/')
 def home():
-    return "Hello world!"
+    return jsonify({"message": "Hello world"}), 200
 
 
 #Get information from a user
@@ -37,7 +50,7 @@ def create_user():
         user_collection.insert_one(user)
     except Exception as e:
         return jsonify({"error": str(e)}), 500
-    return jsonify({"message": f"User '{user}' created successfully"}), 201
+    return jsonify({"message": f"User '{username}' created successfully"}), 201
 
 
 #Create a new quiz and attach it to a user
@@ -63,6 +76,7 @@ def create_quiz():
 #     "choices": ["Blue", "Red", "Yellow", "Green"],
 #     "answer": "Green"
 # }
+#Returns a json object with message and question ID for later referencing
 @app.route("/create-question", methods=["POST"])
 def create_question():
     data = request.get_json()
@@ -73,11 +87,12 @@ def create_question():
     if not list(quiz_collection.find({"quizName": quizName}, {"_id": 0})):
         return f"quiz '{quizName}' not found", 400
     
+    questionID = ObjectId()
     questionObject = {
         "question": data["question"],
         "choices": data["choices"],
         "answer": data["answer"],
-        "_id": ObjectId() #Need to attach an ID so we can modify it later
+        "_id": questionID #Need to attach an ID so we can modify it later
     }
 
     quiz_collection.update_one(
@@ -85,7 +100,7 @@ def create_question():
         {"$push": {"questions": questionObject}}
     )
 
-    return jsonify({"message": f"Question '{data["question"]}' in quiz '{quizName}'created successfully"}), 201
+    return jsonify({"message": f"Question '{data["question"]}' in quiz '{quizName}' created successfully" , "questionID" : str(questionID)}), 201
 
 
 #Gets a quiz and all questions
@@ -94,7 +109,7 @@ def get_quiz(quizName):
 
     quiz = quiz_collection.find_one({"quizName": quizName}, {"_id": 0})
     if not quiz:
-        return "Quiz not found", 404
+        return jsonify({"error": f"Quiz '{quizName}' not found"}), 404
     
     #need to convert objectID objects to string
     for q in quiz["questions"]:
@@ -127,18 +142,22 @@ def delete_quiz():
     data = request.get_json()
 
     quizName = data["quizName"]
+    try:
+        #Find the owner's username
+        quiz = quiz_collection.find_one({"quizName": quizName}, {"_id": 0})
+        if not quiz:
+            return jsonify({"error": f"Quiz '{quizName}' not found"}), 404
+        username = quiz["owner"]
 
-    #Find the owner's username
-    quiz = quiz_collection.find_one({"quizName": quizName}, {"_id": 0})
-    username = quiz["owner"]
+        value = quiz_collection.delete_one({"quizName": quizName})
+        userValue = user_collection.update_one({"username": username}, {"$pull": {"quizzes": quizName}})
 
-    value = quiz_collection.delete_one({"quizName": quizName})
-    userValue = user_collection.update_one({"username": username}, {"$pull": {"quizzes": quizName}})
-
-    if value.deleted_count > 0 and userValue.modified_count > 0:
-        return jsonify({"message": f"quiz '{quizName}'  deleted successfully"}), 200
-    else:
-        return f"Quiz '{quizName}' not found", 404
+        if value.deleted_count > 0 and userValue.modified_count > 0:
+            return jsonify({"message": f"quiz '{quizName}' deleted successfully"}), 200
+        else:
+            return f"Quiz '{quizName}' not found", 404
+    except Exception as e:
+        return jsonify({"error": str(e)}), 500
 
 #Gets the answer for a question
 #Expects JSON object with quizName and questionID
@@ -147,15 +166,15 @@ def check_answer():
     data = request.get_json()
 
     quizName = data["quizName"]
-    questionID = ObjectId(data["questionID"])
+    questionID = ObjectId(data["questionID"]) #Convert id to ObjectID type
 
     question = quiz_collection.find_one({"quizName": quizName, "questions._id": questionID}, {"_id": 0, "questions.$": 1})
 
     if not question:
-        return "No question found", 404
+        return f"Question '{data["questionID"]}' not found", 404
     correctAnswer = question["questions"][0]["answer"]
 
-    return (correctAnswer)
+    return jsonify({"answer" : correctAnswer})
 
 if __name__ == "__main__":
     app.run(host ='0.0.0.0', port = 5000, debug=True)
